Resolve conflict for modelstub
